feat: logout, myRecipe 추가

+++ back/src/main/java/com/gamul/api/request/MyRecipeEditReq.java
@@ -17,7 +17,7 @@ public class MyRecipeEditReq {
     String userName;
 
     @JsonProperty("my_recipe_id")
-    int myRecipeId;
+    Long myRecipeId;
 
     @JsonProperty("image_data_url")
     String imageDataUrl;
